Put BuilderException and builder constants into separate classes
