+++ src/main/java/org/squiddev/luaj/api/builder/tree/LuaClass.java
@@ -2,7 +2,7 @@ package org.squiddev.luaj.api.builder.tree;
 
 import org.squiddev.luaj.api.LuaAPI;
 import org.squiddev.luaj.api.LuaFunction;
-import org.squiddev.luaj.api.builder.APIBuilder;
+import org.squiddev.luaj.api.builder.BuilderException;
 import org.squiddev.luaj.api.builder.BuilderSettings;
 import org.squiddev.luaj.api.transformer.Transformer;
 import org.squiddev.luaj.api.validation.ILuaValidator;
@@ -70,7 +70,7 @@ public class LuaClass {
 			}
 		}
 
-		if (methods.size() == 0) throw new APIBuilder.BuilderException("No @LuaFunction methods", this);
+		if (methods.size() == 0) throw new BuilderException("No @LuaFunction methods", this);
 
 		Set<LuaField> fields = this.fields;
 		for (Field field : klass.getFields()) {
