+++ src/main/java/org/squiddev/luaj/api/builder/tree/LuaMethod.java
@@ -1,7 +1,7 @@
 package org.squiddev.luaj.api.builder.tree;
 
 import org.squiddev.luaj.api.LuaFunction;
-import org.squiddev.luaj.api.builder.APIBuilder;
+import org.squiddev.luaj.api.builder.BuilderException;
 import org.squiddev.luaj.api.builder.Parameter;
 import org.squiddev.luaj.api.validation.ILuaValidator;
 
@@ -118,7 +118,7 @@ public class LuaMethod implements Iterable<LuaArgument> {
 				if (argument.optional) {
 					hasOptional = true;
 				} else if (hasOptional) {
-					throw new APIBuilder.BuilderException("Non-Optional item after an optional one", LuaMethod.this);
+					throw new BuilderException("Non-Optional item after an optional one", LuaMethod.this);
 				} else {
 					length++;
 				}
