+++ src/main/java/org/squiddev/luaj/api/setters/LoaderSetter.java
@@ -1,12 +1,13 @@
 package org.squiddev.luaj.api.setters;
 
 import org.objectweb.asm.MethodVisitor;
-import org.squiddev.luaj.api.builder.APIBuilder;
 import org.squiddev.luaj.api.builder.APIClassLoader;
+import org.squiddev.luaj.api.builder.BuilderException;
 import org.squiddev.luaj.api.builder.IInjector;
 import org.squiddev.luaj.api.builder.tree.LuaField;
 
 import static org.objectweb.asm.Opcodes.GETSTATIC;
+import static org.squiddev.luaj.api.builder.BuilderConstants.CLASS_LOADER;
 
 /**
  * Sets the field to be the class loader
@@ -21,9 +22,9 @@ public class LoaderSetter implements IInjector<LuaField> {
 	@Override
 	public void inject(MethodVisitor visitor, LuaField field) {
 		if (!field.field.getType().isAssignableFrom(APIClassLoader.class)) {
-			throw new APIBuilder.BuilderException("Cannot convert " + field.field.getType().getName() + " to APIClassLoader", field);
+			throw new BuilderException("Cannot convert " + field.field.getType().getName() + " to APIClassLoader", field);
 		}
 
-		visitor.visitFieldInsn(GETSTATIC, field.klass.name, "LOADER", APIBuilder.CLASS_LOADER);
+		visitor.visitFieldInsn(GETSTATIC, field.klass.name, "LOADER", CLASS_LOADER);
 	}
 }
